% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cegs.R
\name{cegs}
\alias{cegs}
\alias{cegs_ml}
\alias{cegs_ld}
\title{CEGS Estimators (ML and LD)}
\usage{
cegs_ml(n)

cegs_ld(n)
}
\arguments{
\item{n}{Integer vector of species counts.}
}
\value{
A list with fields \code{richness}, \code{scale}, \code{shape}, \code{AICc}, \code{fitted_RAD}, \code{fitted_SAD}.
}
\description{
Maximum likelihood (ML) and likelihood-density (LD) estimators for the CEGS model.
}
