% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_cegs.R
\name{nearest_neighbour}
\alias{nearest_neighbour}
\title{Euclidean nearest neighbour (on complete rows only)}
\usage{
nearest_neighbour(M)
}
\description{
Euclidean nearest neighbour (on complete rows only)
}
\keyword{internal}
