% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_ecoregister.R
\name{summarise_per_inventory}
\alias{summarise_per_inventory}
\title{Summarise per-inventory richness and individuals}
\usage{
summarise_per_inventory(x)
}
\arguments{
\item{x}{A data.table as returned by \code{\link[=read_ecoregister_counts]{read_ecoregister_counts()}}.}
}
\value{
A data.table with columns \code{sample_id}, \code{S}, \code{N}.
}
\description{
Summarise per-inventory richness and individuals
}
