% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_cegs.R
\name{tail_histogram}
\alias{tail_histogram}
\title{Build tail frequency-of-frequencies for a neighbour j}
\usage{
tail_histogram(counts, k = 10L, max_x = 4096L)
}
\description{
Build tail frequency-of-frequencies for a neighbour j
}
\keyword{internal}
