% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_cegs.R
\name{tail_nll}
\alias{tail_nll}
\title{Negative log-likelihood of neighbour tail under fitted SAD p (conditional on presence)}
\usage{
tail_nll(s, u, p)
}
\description{
Negative log-likelihood of neighbour tail under fitted SAD p (conditional on presence)
}
\keyword{internal}
